Add DishDetailsActivity
Added new activity which can be navigated from dishes list fragment
by short clicking any dish item. It shows dish details: name and
ingredients and also has two buttons: 'Save' and 'Remove' which are
non functional for now. Also note: ingredients are not yet implemented.

One thing to consider is how the dish name is displayed. Currently it is
received as part of an intent starting the activity,
which may not be ideal.