Display ingredients of a dish in dish details view
*DishIngredient and DishIngredientDao - Entity and Data Access Object
for dishes_ingredients_table which is a associative table allowing
many-to-many relationship between dishes_table and ingredients_table

*Ingredient and IngredientDao - Entity and Data Access Object for
ingredients_table which contains ingredients used in the app

Note: Ingredients are displayed using TextView object in
DishDetailsActivity. This is most certainly not the best solution and
will be changed in the future.